refactor: 폴더 이동(CommunityPostTypographyButton)
CommunityPost -> CommunityPostSection
